Posts on Command Bar :)

+++ graphql/resolvers/mutations/user/index.ts
@@ -1,5 +1,6 @@
-import { Context } from '../../../context'
+import { Context, SafeContext } from '../../../context'
 import { Role } from '../../../types.generated'
+import { AuthenticationError } from 'apollo-server-nextjs'
 
 export async function upsertUser(
   _: any,
@@ -56,12 +57,27 @@ export async function upsertUser(
   return user
 }
 
-export async function deleteUser(_: any, { id }: { id: string }, ctx: Context) {
-  const { prisma } = ctx
-  const user = await prisma.user.delete({
+export async function deleteUser(
+  _: any,
+  { id }: { id: string },
+  ctx: SafeContext
+) {
+  const { prisma, viewer } = ctx
+  const user = await prisma.user.findUnique({
     where: {
       id,
     },
   })
-  return user
+
+  if (!user) return true
+
+  if (user.id !== viewer.id && !viewer.isAdmin) {
+    throw new AuthenticationError('You are not allowed to delete this comment')
+  }
+
+  await prisma.user.delete({
+    where: {
+      id: id,
+    },
+  })
 }
